PlotAll {rawDiag}	R Documentation

PlotAll applies all available rawDiag plot functions to a given rawDiag object

Description

PlotAll applies all available rawDiag plot functions to a given rawDiag object

Usage

PlotAll(x, prefix = "primer", savepng = TRUE, resolution = 240)

Arguments

x
a data.frame fullfilling the is.rawDiag column naming criteria.

prefix
filename prefix if plots a writting to a png file

savepng
default is TRUE

resolution
number of pixels in png file, default is 240

Author(s)

Christian Panse <cp@fgcz.ethz.ch>, 2018

References

Table 1 in https://doi.org/10.1101/304485

Examples

data(WU163763)
WU <- WU163763[WU163763$filename %in% unique(WU163763$filename)[1:2], ]
rv <- rawDiag:::PlotAll(x = WU, savepng = FALSE)

[Package rawDiag version 0.0.10 Index]