as.rawDiag {rawDiag}	R Documentation

Force an Object to Belong to a Class rawDiag

Description

Force an Object to Belong to a Class rawDiag

Usage

as.rawDiag(object)

Arguments

object
any R object.

Value

a data.frame having the column names defined in rawDiag:::.getDesiredColNames.

Author(s)

Christian Panse <cp@fgcz.ethz.ch>, 2018

[Package rawDiag version 0.0.10 Index]