benchmark {rawDiag}	R Documentation

benchmark

Description

a data set for benchmarking the throuput of read.raw and as.rawDiag.mzR functions.

Usage

data("benchmark")

Format

Three data frames.

b.Linux

performed benchmark_raw on a Linux server

b.Apple

on a apple laptop

X.Linux

performed rawDiag:::benchmark_mzR on a Linux server

Details

As benchmark data, we downloaded the raw files described in pmid29183128 on the filesystem availbale through proteomeXchange PXD006932. For the benchmark we limited the input to 128 files, corresponding to two times the available number of processor cores of the Linux system. For comparison to the open file standards using the mzR package mzR, we converted the raw files into mzML files using version ProteoWizard 3.0.11252 64-bit on Windows 10.

Source

the benchmark was performed on the following public available data proteomeXchange PXD006932.

References

http://planetorbitrap.com/rawfilereader#.WyThDK3QPmE

See Also

benchmark_raw

Examples

data(benchmark)
length(unique(b.Linux$nrow))
round(sum(unique(b.Linux$file.size)) / 1024^3)
format(sum(unique(b.Linux$nrow)), scientific=FALSE, big.mark = "'")
rawDiag:::.technote_benchmark_figure_1()
rawDiag:::.technote_benchmark_figure_2()

[Package rawDiag version 0.0.10 Index]