read.raw {rawDiag}	R Documentation

mass spec reader function

Description

this function is a generic adapter function for reading mass spectrometric measurement. It requieres the New RawFileReader dll files

ThermoFisher.CommonCore.BackgroundSubtraction.dll,

ThermoFisher.CommonCore.MassPrecisionEstimator.dll

ThermoFisher.CommonCore.RawFileReader.dll

be installed on the system. This can be done by using the Global Assembly Cache (GAC) or by setting the MONO_PATH enviroment.

Usage

read.raw(file, mono = if (Sys.info()["sysname"] %in% c("Darwin", "Linux"))
  TRUE else FALSE, exe = file.path(path.package(package = "rawDiag"), "exec",
  "fgcz_raw.exe"), mono_path = "", rawDiag = TRUE, argv = "qc",
  system2_call = if (Sys.info()["sysname"] == "Windows") TRUE else FALSE,
  method = "thermo", ssh = FALSE)

Arguments

file
the name of the Thermo Fisher Scietific raw file which the data are to be read from.

mono
a boolean indicates if the mono enviroment should be used.

exe
the adapter software; default is set to file.path(path.package(package = "rawDiag"), "exec/fgcz_raw.exe")

mono_path
define the PATH where the ThermoFisher.CommonCore* dll files are located; default is "" (empty string) ; no MONO_PATH is set.

rawDiag
boolean; if TRUE (default) the output is coerced to an rawDiag object. if FALSE the function will return all available columns.

argv
argument vector for the adapter.

system2_call
calles system2 if TRUE; otherwise it uses a pipe.

method
the adapter type. default is 'thermo' which read Thermo Fisher Scietific raw files.

ssh
boolean, if ssh pipe should be used. defaul is FALSE

Value

a data.frame.

Author(s)

Christian Panse <cp@fgcz.ethz.ch>, 2017, 2018

References

http://planetorbitrap.com/rawfilereader

http://www.mono-project.com/docs/advanced/assemblies-and-the-gac/

Examples

(rawfile <- file.path(path.package(package = 'rawDiag'), 'extdata', 'sample.raw'))
system.time(RAW <- read.raw(file = rawfile))
dim(RAW)
summary(RAW)
PlotScanFrequency(RAW)

# read all dimensions
dim(RAW)
RAW <- read.raw(file = rawfile, rawDiag = FALSE)
dim(RAW)

[Package rawDiag version 0.0.10 Index]